Выполнил восьмое задание - не до конца
