--- modules/getRandom.js
@@ -1,1 +1,0 @@
-export const getRandom = (meaning) => Math.ceil(Math.random() * meaning);
