--- modules/getTime.js
@@ -1,6 +1,0 @@
-function getTime () {
-    const time = new Date;
-    return `${time.getHours()}:${time.getMinutes()}`;
-};
-
-export default getTime;
